Add external test for RFC2833 DTMF detection.
From reviewboard:
There are three test cases run. In each, SIPp is used to place a call to Asterisk, playback one or multiple pcap files and then send a BYE. The test.lua script is set to run a handler for each VarSet manager event that is received. In the handler, we compare the received DTMF digits to what we expected to receive and fail if there is a discrepancy.

The pcap files that I used for playing DTMF could not be attached, so I will include a brief description of them here:

dtmf_2833_*.pcap: These are the RFC2833 DTMF pcap files included in the distribution of SIPp. Strangely, they do not include a '0' DTMF digit.
dtmf_2833_*_noend.pcap: These are generated from the original DTMF pcap files by chopping off the end frames.
broken_dtmf.pcap: This is a series of digits that was collected from the reporter of Mantis issue 15811 who was having a problem with duplicate DTMF being detected. The quirk with the DTMF here is that there are out-of-order frames received.

If more information is required regarding the tests run, see
tests/rfc2833_dtmf_detect/test-config.yaml

Review: https://reviewboard.asterisk.org/r/560



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@143 a5964663-950a-4945-9369-5dd0b5cb08c1