Add new python module to facilitate monitoring manager events when testing.
A basic sample has been provided in the run-test file. More documentation
regarding use is present in the client module in /lib/python.

Some notes:
Since the main purpose of this is to be used by other tests, the ami-monitor
test has not been added to the tests.yaml file.
There is some quirky code present that makes things work properly with the
paths that allows for direct execution of the test, which I call standalone
mode. Obviously other tests do not need to copy this behavior.

Review: https://reviewboard.asterisk.org/r/637/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@305 a5964663-950a-4945-9369-5dd0b5cb08c1