Add "sendfullybooted" option to default asterisk.conf file.
Without this, tests that run on non-trunk branches will not hang forever
due to their inability to catch the FullyBooted event.

A further modification would be to make use of the CLI command in the
asterisk:spawn() function so that it will not return until fully booted.
Then there would be no need to have individual tests use an AMI connection
for that purpose.



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@353 a5964663-950a-4945-9369-5dd0b5cb08c1