Make wrapup_time test a bit more reliable.
Bamboo was reporting test failures. The biggest reason for
this was that we were only waiting for a call to be answered before
moving on to the next one. What we needed to be doing was to wait
for the previous call to complete before moving to the next one. After
making these changes, I find that the Bamboo agent is passing the
test every time now.

One more errant test to go...



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@368 a5964663-950a-4945-9369-5dd0b5cb08c1