New install_configs() function and ability for backwards compatible configs
Add the ability to specify a config directory to be installed.  Everything within
the directory, except the .svn, branch-1.4, branch-1.6.2 and branch-1.8
directories, will be copied over.

Also includes in functionality for backwards compatible config files.  If a
sub-directory is found to be the same as the running instance, it will use that
version of .config file.

For example: If testsuite is running a test against 1.4 (branch-1.4):

configs/manager.conf
configs/sip.conf
configs/branch-1.4/sip.conf

Because the sip.conf file exists in the branch-1.4 directory, it will be used
in place of the top-level sip.conf.  As for the manager.conf file, because it
does not exists in the branch-1.4 direcory, the top-level manager.conf will
be used. 

Review: https://reviewboard.asterisk.org/r/831/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@680 a5964663-950a-4945-9369-5dd0b5cb08c1