Pass DTMF between Asterisk instance before starting tests.
The current fastagi tests are dependent on the speed of the host computer, on a
slow system they will usually fail. The 2nd reactor has not full started but the
1st has already started the test.  I've now added 2-way communication between
the dialplans.  Since side A starts first, we now WaitExten(10) expecting side B
to send DTMF (it confirms B is ready).

I also noticed we don't properly stop reactor A, which was an issue on the
FreeBSD host.  It basically created a deadlock on 'core stop now', eventually
the testsuite had to kill the process to stop asterisk.

Review: https://reviewboard.asterisk.org/r/1035/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@1011 a5964663-950a-4945-9369-5dd0b5cb08c1