Remove OptionParser from tests
This patch removes the ability to pass command-line arguements to test. Not only
is this a sloppy way of doing things, 95% of the time is was not even required.
Because some tests used the ast_version from the argv parameter, we've added the
cli_aliases.conf file to help with CLI syntax.

Additionall, I've removed valgrid support (was not used for anything) until we
decided how to find a better solution for it.  Ideally, the testsuite will have
a .config file all tests can access for global settings.

Review: https://reviewboard.asterisk.org/r/1386/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@1962 a5964663-950a-4945-9369-5dd0b5cb08c1