Added pre- and post-test condition check controller to testsuite
This adds the ability to define pre- and post-test conditions to be checked
before and after a test is executed.  As an example of this, an Asterisk thread
check is included with this patch.  This patch also includes the following:
* the ability to define subsets of tests for execution via the --test option, 
as opposed to a single test.
* the ability to define build options as a Dependency in the test-config.yaml

(closes issue ASTERISK-18430, ASTERISK-18424, ASTERISK-18428)
Review: https://reviewboard.asterisk.org/r/1430/

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@2210 a5964663-950a-4945-9369-5dd0b5cb08c1