Add SIP / Incomplete handling tests
This adds a test to check for address incomplete handling
for SIP and the Incomplete dialplan application.  These tests
check that the proper hangup code is set in the correct scenarios
and that the dialplan handles waiting for additional DTMF tones
in the appropriate fashion in the presence of chan_sip.

Review: https://reviewboard.asterisk.org/r/1417/

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@2357 a5964663-950a-4945-9369-5dd0b5cb08c1