Adds a new test channels/SIP/sip_tls_register which uses sipp to register a SIP account.
The test checks both Asterisk's manager events to see if the registration resulted in an
appropriate change in the registration state of the peer and sipp's return value for
whether or not the SIP exchange was appropriate to sipp's expectations. This revision
also adds a function to utils for replacing strings within files, which can help with
modifying config files to use directories created within Asterisk instance folders.

Review: https://reviewboard.asterisk.org/r/1548/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@2860 a5964663-950a-4945-9369-5dd0b5cb08c1