Adds tests for 'queue set penalty' command and analogous AMI command
This test basically just has a few preset queue members in a number of queues and the
test is set up to test different uses of the command as a state machine. It then checks
QueuePenalty events sent in response to the commands and makes sure they are in line with
expectations. This test was made in response to a regression that was introduced when
adding realtime support to queue set penalty.

Review: https://reviewboard.asterisk.org/r/1611/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@2907 a5964663-950a-4945-9369-5dd0b5cb08c1