Update asttest's version parsing
Asterisk no longer provides a header file with the version defined in it.
The previous mechanism of parsing this file no longer work.  Previously,
in r3021, the python library's mechanism of parsing this was updated
to obtain the version from the installed Asterisk binary; this patch
adds that functionality to asttest's lua library as well.  It also addresses
some minor points from the python review.  See
https://reviewboard.asterisk.org/r/1695/

Review: https://reviewboard.asterisk.org/r/1701/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3022 a5964663-950a-4945-9369-5dd0b5cb08c1