Make SIP blind transfer tests tolerant of multiple channel status callback results
When checking for the success of a blind transfer, multiple results can be
returned, depending on the timing of the test.  The tests have now been changed
to iterate over the results checking each for the expected bridged channel, and
will only fail the test if no results contain the expected results.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3065 a5964663-950a-4945-9369-5dd0b5cb08c1