Perform initial tagging of tests
This patch builds on r2724 by adding tags to the vast majority of tests.
In addition, it adds a new command line parameter, '-L', which lists
the available tags in the Test Suite.  A tag is counted as being 'available'
if there exists at least one test that will be executed if that tag parameter
is the only tag parameter passed to runtests.

Review: https://reviewboard.asterisk.org/r/1791/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3077 a5964663-950a-4945-9369-5dd0b5cb08c1