Migrate two fastagi tests to use twisted for process management
Tests fastagi/execute and fastagi/wait-for-digit were missed in the
migration patch, due to the fact that neither actually failed.  In
the case of wait-for-digit, that's because its not executed; execute
did not fail despite the fact that Asterisk never started.  This patch
migrates them to use twisted for their process management in the same
manner as the other non-TestCase derived fastagi tests.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3135 a5964663-950a-4945-9369-5dd0b5cb08c1