Make ami_login_error return a twisted Failure
Per the specification for deferds, an errback must either raise
an exception or return a twisted.python.failure.Failure instance.
Otherwise, the usual chain of callbacks is returned to.  This isn't
terribly useful when the error in question is a failure to connect
to AMI.  This should make the login error a bit more terminal when
something goes wrong during an AMI connection.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3152 a5964663-950a-4945-9369-5dd0b5cb08c1