Update the pre-/post-condition check framework to support twisted
This patch updates the pre-/post-condition checking framework in the
Asterisk Test Suite to use the twisted asynchronous framework for its
CLI/AMI commands.  This was necessary, as the starting/stopping
of Asterisk now expects things that explicitly need to be involved
in the process to be in the deferred callback chain.

Review: https://reviewboard.asterisk.org/r/1831

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3188 a5964663-950a-4945-9369-5dd0b5cb08c1