Add ConfBridge Recording Test
This test checks for recording of a conference in ConfBridge via
the MixMonitor application.  It tests enabling and setting the location
of the recording via the CONFBRIDGE function and via the confbridge.conf
configuration file.  After a conference is recorded, the test verifies
that the recorded file contains some audio via the BackgroundDetect
application.

Review: https://reviewboard.asterisk.org/r/1880/

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3213 a5964663-950a-4945-9369-5dd0b5cb08c1