Fix the sip_cause test
It appears that the dialplan post-Dial is executing faster than the
SIP_CAUSE hash can be set via the MASTER_CHANNEL() dialplan function.
This is causing the user event with cause information to be reported
before the final value is set into SIP_CAUSE. The Wait(2) (gross, I
know) should work around this problem.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3227 a5964663-950a-4945-9369-5dd0b5cb08c1