Disable masquerade test until failures are debugged
It appears as if having the timing source be timer_fd is causing
local channel optimizations to hang.  This does not appear to be the
case with DAHDI as the timing source - so until this can be debugged
further, the test is being disabled.  ASTERISK-19966 created to track
and resolve this problem.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3260 a5964663-950a-4945-9369-5dd0b5cb08c1