Update play_message test to be tolerant of starpy
While starpy in our github repo should be attempting to use ',' to
delineate channel variables, for whatever reason, whatever version of
starpy is installed is using '|' characters.  Rather then rely on starpy
to update its nomenclature to a supported version of Asterisk, we simply
pack the channel variables to set in an Originate into one long string.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3279 a5964663-950a-4945-9369-5dd0b5cb08c1