Tweak Asterisk start cycle to be more tolerant of config installation
On some machines, installing the configs - which consists of copying a
number of Asterisk configuration files and making a large number of
symbolic links - can take the OS longer to complete then what the method
execution time infers.  As such, Asterisk can sometimes be started before
the test specific Asterisk configuration file has been created, which
can lead to Asterisk using the default configuration information.  This
causes a problem for the test, which - for all CLI commands and other
interactions - will use the one installed in /tmp/asterisk-testsuite/.
We now defer starting Asterisk for a second, to give time for the config
installation to complete.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3305 a5964663-950a-4945-9369-5dd0b5cb08c1