Restore SIP TCP attended transfer test to functionality
This test had stopped working quite a while ago and had been set to be
skipped on all test runs because we did not have time to fix it.  The
test has been revamped in much the same manner as the
sip_attended_transfer test and now functions properly.

(issue SWP-4660)


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3442 a5964663-950a-4945-9369-5dd0b5cb08c1