Add CDR/CEL Queue tests for No Answer scenarios
These tests cover CDR/CEL behavior defined for Queues in Asterisk 12 (see 
r375416).  In that code change, when a single Queue member was paused, if
the caller failed to be answered by another Queue member the overall
disposition was BUSY.  That patch changed the behavior to be NO ANSWER.
These tests cover the CDR/CEL behavior in the following three scenarios:
 * When no one answers, and no member is 'busy'
 * When no one answers, and a single member is 'busy' but at least one other
   member could have picked up the phone (lazy gits)
 * When no one answers, and all members were legitimately 'busy' (your call
   is important to us, but so is our lunch break kkthnx)

Hopefully, if someone goes in and mucks with CDRs in the future, we'll at
least get a big red flag that someone changed the behavior on us... again.

Review: https://reviewboard.asterisk.org/r/2063/

(issue AST-906)
Reported by: Thomas Arimont



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3501 a5964663-950a-4945-9369-5dd0b5cb08c1