Add provisions to allow additional extensions.conf logic for bridge tests that require it.
The extensions.conf files for "Alice" and "Bob" in bridge tests now include a file
called "extensions-extra.conf". A blank version of this file is included in the central
configuration location. If a specific test requires additional global variables, extensions,
or contexts, then an extensions-extra.conf file can be added into the specific test in
order to fulfill specific needs.

This will be useful for tests such as connected line update tests, where Alice and Bob
will need to alert the test that they received connected line updates. This could also
be useful for transfer tests. You never know.



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3577 a5964663-950a-4945-9369-5dd0b5cb08c1