Add AMI event triggering of originations
This adds the ability to trigger originations from any AMI event using
YAML configuration as an extension of the Originator pluggable module.
As a part of this enhancement, AMIPrivateCallbackInstance was added
which can be used to hook AMI events with an arbitrary callback if a
YAML definition of the event is already available in the format used
by AMIEventModule.

Review: https://reviewboard.asterisk.org/r/2261/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3606 a5964663-950a-4945-9369-5dd0b5cb08c1