Add tests to cover RFC 3327 support in chan_sip
This patch adds a test for the Path header support in chan_sip. It covers two
scenarios:

1. A UA sends a REGISTER request to Asterisk. The test verifies that the Path
header in the request is copied into the 200 OK response and that the response
indicates support for 'path'. It then verifies that a subsequent request sent
to the UA (An INVITE for a new call) contains the path information in the
Route header.

2. If path support is enabled globally, it verifies that outbound REGISTER
requests from Asterisk indicate support for 'path'.

Review: https://reviewboard.asterisk.org/r/2325

(issue ASTERISK-16884)



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3658 a5964663-950a-4945-9369-5dd0b5cb08c1