Make application tests more tolerant to different event ordering
The apptest tests use a Local channel (which is a pair of channels) to drive
the logic within a complex Asterisk application. When a test starts, the
apptest module must determine which half of the Local channel is in the
application under test, and which half is in a set of extensions that drive
input into the application.

Prior to some Asterisk refactoring of AMI events on Stasis-Core, the test
would always receive the Newchannel events for the Local channel halves
prior to receiving the Varset event notifying it that the Local channel is
the channel the test explicitly spawned. Now, however, that ordering does not
always occur - in fact, it was never guaranteed to happen (it just did!)
This patch makes the apptest tests more tolerant to the Varset event
occurring before or after the second Newchannel event.



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3684 a5964663-950a-4945-9369-5dd0b5cb08c1