Add pluggable CEL verification module
This uses AMI to allow verification of CEL messages. It can be configured to
allow CEL messages to come before and/or after others if they are not always
in the same order due to channels operating on their own threads of execution.

(issue ASTERISK-21422)
(closes issue ASTERISK-21422)

Review: https://reviewboard.asterisk.org/r/2498/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3783 a5964663-950a-4945-9369-5dd0b5cb08c1