Don't create a race condition between hanging up ast1 and ast2
In the SIP TLS test, Asterisk instance 1 sends a DTMF tone to Asterisk
instance 2 and waits for the second instance to break it out of a
Background application. Asterisk instance 2 sends a DTMF and then hangs
up. The problem is the hangup of Asterisk instance 2 may cause instance
1 to hangup prior to servicing the DTMF. This race condition always
existed; however, with the DTMF/AMI changes, it's much more likely to hit
it.

Now only the last instance calls Hangup. The other instance will Echo until
hung up.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3791 a5964663-950a-4945-9369-5dd0b5cb08c1