Add a hangup event instance class to PluggableModules
It's a fairly common need to arbitrarily hang up channel. This patch
adds two modules to do this. The first will hangup a channel given an
AMI event that contains the channel to hang up. The second is less picky -
it tracks all channels and, upon matching on some arbitrary event - hangs
up everyone.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3860 a5964663-950a-4945-9369-5dd0b5cb08c1