Add module for testing RESTful API's
This patch adds a pluggable module to the Asterisk testing framework
to aid in testing the new Asterisk 12 RESTful API's.

It is similar in scope to the AMI plugable module. In the test's
test-config.yaml file, event specifications can be specified, along
with requirements for passing the test. The only requirements
implemented by this patch are count (which support >, < and exact
count) and callback (which is also where custom logic may be kicked
off on events).

The WebSocket client is implemented using autobahn's WebSocket
support[1], since that plays very nicely with our Twisted infected
framework. RESTful API invocations are made using Requests[2], since
it's an extremely convenient abstraction for HTTP.

A test for the /channels/{channelId}/continue resource was added for
the initial implementation.

 [1]: http://autobahn.ws/
 [2]: http://docs.python-requests.org/en/latest/

(closes issue ASTERISK-21818)
Review: https://reviewboard.asterisk.org/r/2618/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3866 a5964663-950a-4945-9369-5dd0b5cb08c1