Print out the test pass/fail status as tests complete
While the final test summary will list out all of the pass/fail statuses of all
tests, it is sometimes useful to know as tests are running if they passed or
failed. Sometimes this can be gleaned from WARNING/ERROR messages coming from
the running process; sometimes, however, processes don't display much when
they fail.

This adds a single line of output to display if the test passed or failed.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3995 a5964663-950a-4945-9369-5dd0b5cb08c1