Don't raise an exception if csvreader can't be created
If csvreader can't be created, the AsteriskCSV class will already handle it
gracefully. Raising an exception creates a situation in which a twisted errback
may have to handle the error while the reactor is shutting down. While an
errback could be created to handle this, the complexity with the Asterisk
shutdown sequence tends to result in situations where it is harder to handle
the exception (and thus prevent the entire test from hanging) than it is
to simply not have an exception be raised and let the CSV processing fail.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@3998 a5964663-950a-4945-9369-5dd0b5cb08c1