Allow many of the bridging tests to properly run against Asterisk 11
This changes several of the bridging tests that were originally developed for
Asterisk 12 so that they are able to run against Asterisk 11.

Additionally the TEST_FRAMEWORK dependency was added to tests that use the
feature configuration of BridgeTestCase as it depends on the test events.

Review: https://reviewboard.asterisk.org/r/2750/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4026 a5964663-950a-4945-9369-5dd0b5cb08c1