Give ARI its own test suite object
Testing ARI is a bit different than testing normal Asterisk tests. Typically, a
test in the Asterisk Test Suite uses the following model:

1. Create Asterisk Instances
2. Wait for instances to fully boot
3. Connect AMI
4. Wait for AMI connection
5. Spawn channels for test

Most of the test logic occurs in step number 5 - while it's possible to inject
logic into the first four steps without writing a new test object, it isn't
easy, particularly if what you need to do is defer the execution of one of
those four steps until something else happens. In particular, ARI tests need
to connect the websocket prior to step 4, as the AMI connection is what
typically causes channels to be created in the dialplan and test execution to
begin. If the websocket connection doesn't happen before that point, a channel
may enter into the Stasis application before the websocket connects and the
test will fail.

This patch gives ARI its own test object so that it can properly connect the
websocket prior to the AMI connection. This changes the startup routine to be:

1. Create Asterisk Instances
2. Wait for instances to fully boot
3. Connect ARI
4. Wait for ARI connection
5. Connection AMI
6. Wait for AMI connection
7. Spawn channels for test

Review: https://reviewboard.asterisk.org/r/2661/

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4065 a5964663-950a-4945-9369-5dd0b5cb08c1