testsuite: Add a test for reinvite to T.38 with directmedia
This is a test that sets up a SIP call from one SIPp endpoint to
another with directmedia with just audio and then a reinvite
is issued from one of the peers to the other to change the call
into a T.38 fax session. The test ensures that at this point,
media is routed back through Asterisk.

(closes issue ASTERISK-17273)
Reported by: dario
Review: https://reviewboard.asterisk.org/r/2854/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4213 a5964663-950a-4945-9369-5dd0b5cb08c1