Provide a configuration for the Originator module in the SIP path test
Prior to this patch, the SIP path test used the PluggableModule.Originator
module with a default configuration. Unfortunately, this caused it to
originate 3 calls when only a single call is expected by sipp. This caused
the test to timeout and kill asterisk, though it still reported success.

After this patch, we only originate the single call that the test expects.

(closes issue ASTERISK-22576)
Reported by: Corey Farrell
patches:
  testsuite-path-config.patch uploaded by coreyfarrell (License 5909)



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4234 a5964663-950a-4945-9369-5dd0b5cb08c1