Update cdr_fork_end_time last application testing
Because the forking of a CDR occurs on a different thread of execution than
the actual recording of the state of the channels that comprise that CDR, it is
possible for the last application to be recorded not as "ForkCDR" but as
something else. Don't bother checking for the last application.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4295 a5964663-950a-4945-9369-5dd0b5cb08c1