Delay slightly when sending the DTMF to cancel the parallel dial attempts
If the DTMF is sent too fast, the hangup cause code frame might not have been
enqueued yet on the channels. This results in a lack of hangup cause information
which can cause the tests to fail. This updates the python code that sends the
DTMF to cancel the test to wait just slightly after the beginning of the dial
attempt to give the channel driver some time to catch up.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4298 a5964663-950a-4945-9369-5dd0b5cb08c1