Testsuite: pjsip_messaging in dialog test sporadically failing
The pjsip message_in_dialog test would fail every once in a while because the
SIPp scenarios were not giving enough time for the call to properly initialize,
send the message and receive a response before hanging up.  The solution was
to have the message receiver send a "bye" (instead of the original sender) once
the message was received.

(closes issue ASTERISK-22777)
Reported by: Matt Jordan


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4308 a5964663-950a-4945-9369-5dd0b5cb08c1