pjsip tests: interactions with chan_sip cause test failures
While chan_sip and chan_pjsip can be loaded and used at the same time it was
causing problems in the testsuite because their defaults try to bind to the
same host:port.  So for the testsuite made is so any test that depends on
res_pjsip will "noload" chan_sip.  Instead of manipulating every current
res_pjsip test (through their own modules.conf or config setting), this fix
checks to see if the test requires/depends on res_pjsip and if it does it
appends a noload => chan_sip.so to the copied modules.conf for that test.

This functionality is configured via the newly added conflicts.txt file.
Any new conflicts that may arise hopefully can reuse the mechanism.

(closes issue ASTERISK-22887)
Reported by: Matt Jordan
Review: https://reviewboard.asterisk.org/r/3060/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4454 a5964663-950a-4945-9369-5dd0b5cb08c1