chanspy_w_mixmonitor: abandon Console channel driver; update to TestCase
This patch updates the chanspy_w_mixmonitor test such that it no longer uses
the console channel driver. It now uses a Local channel attached to a MoH
source to test the ChanSpy-ing. Additionally, it is now based on the TestCase
class, such that it gains the benefits of using that common base class.



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4462 a5964663-950a-4945-9369-5dd0b5cb08c1