PEP 8-ify the Asterisk Testsuite python libraries
This patch does a reasonable job of making the Python code in
lib/python/asterisk compliant with PEP 8. (Reasonable being > 9 on the
pylint "how am I doing?" code rating)

A few disclaimers:
(1) This patch does not go into the python code for the tests themselves, other
    than to make the tests run. Making the individual tests PEP 8 compliant
    would be nice, but it's a whole lot more work with less benefit.
(2) Not *everything* was cleaned up in the libraries. In general, this patch
    tries to get the vast majority of findings, but there are a few minor
    nitpicks here and there.
(3) This patch is scarier than it looks. No semantics (hopefully!) were changed
    in the process of doing this: this is all just formatting/guidelines/etc.

Note that this patch does break the world as it renames a number of python
modules, functions, etc.

Next time we'll pay a bit more attention to style when we start these kinds of
projects.... really. I promise.

Review: https://reviewboard.asterisk.org/r/3096/




git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4556 a5964663-950a-4945-9369-5dd0b5cb08c1