confbridge: Add tests for waitmarked/marked user interaction
This patch adds two tests:

confbridge_waitmarked_single: This test is a simple Marked/WaitMarked test that
verifies basic interaction between Marked and WaitMarked users. In this test,
three scenarios are executed:
(1) A scenario that tests that when a WaitMarked user enters first and leaves
    last, the conference puts them into an INACTIVE state
(2) A scenario that tests that when a Marked user enters first the conference
    starts, but when the Marked user leaves the WaitMarked user is put into an
    INACTIVE state
(3) A scenario that tests that a Marked user always keeps the conference
    active, regardless of a WaitMarked user entering/leaving

confbridge_waitmarked_kick: This test verifies the bug that was fixed in
r405217. In it, three WaitMarked users join the ConfBridge, followed by a
Marked user. The Marked user is kicked, and then immediately during the user
being eject, the three WaitMarked users are also kicked. This test verifies
that the ConfBridge conference correctly transitions through the states even
in the presence of multiple users of different types being ejected.

Review: https://reviewboard.asterisk.org/r/3121/

(issue AST-1258)
Reported by: Steve Pitts


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4579 a5964663-950a-4945-9369-5dd0b5cb08c1