feature_call_pickup: Correct channel variable usage in check_bridge_enter
In the handler for the BridgeEnter event, the 'channel2' variable was still
being used to  reference the pickup channel to hangup. In Asterisk 12,
this should have just been 'channel'. Note that the test still passed even
with this error; it would just time-out. This change now lets the test
run to completion and terminate itself appropriately.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4651 a5964663-950a-4945-9369-5dd0b5cb08c1