pbx/ast_db: Add tests for the DB family of functions
This test verifies that the following AstDB functions and variables work as
intended - DB, DB_RESULT, DB_EXISTS, DB_KEYS, DB_DELETE.

The first iteration tests input and output using DB(family/key) as well as
the DB_RESULT variable.

The second iteration tests DB_EXISTS(family/key) with a non-existant and
existant database, and also tests DB_KEYS(family) with a non-existant
database, a family with one key, and a family with multiple keys.

The last iteration tests DB_DELETE(family/key) on a non-existant database
and an existant database. If it is called when there is nothing to delete,
it will return an empty string.

Note that this patch was written by Ben Ford.

Review: https://reviewboard.asterisk.org/r/3220/

(closes issue ASTERISK-23187)



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4707 a5964663-950a-4945-9369-5dd0b5cb08c1