testsuite: LinkedID Propagation test
This monitors bridge enter/exit and CEL events, and builds
a picture of which channels are bridged together, and when
their LinkedID changes, and then verifies that each change
is correct.

Note: The BridgeEnter and CEL 'BRIDGE_ENTER' events can be
in either order - so wait for both to arrive before making
a judgement on the LinkedID.

This test can be added to any other bridging test to add a
layer of LinkedId Propagation checking to it.

The bridge_action test has been changed to use an event to
trigger channel creation sequence which resolves a problem
with creation order determination.

(issue ASTERISK-23120)
Review: https://reviewboard.asterisk.org/r/3182/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4723 a5964663-950a-4945-9369-5dd0b5cb08c1