originate-cdr-disposition: Fix test failure
This modifies the expected originate success condition for iterations 2 and 3.
In these iterations, the Dial operation completes with either a result of
"Congestion" or a result of "Failed". Since an Originate only succeeds if the
channel is Answered, it should be expected that these two operations would
fail as well.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4754 a5964663-950a-4945-9369-5dd0b5cb08c1