testsuite: eliminate version check zombies
While running the testsuite on a virtual machine, I noticed that
it would occasionally fail to start, complaining of being unable
to kill sipp (code introduced svn r4677 and rb 3198).  I tracked
this back to an instance of sipp that was run to get the version
number and was left as a zombie. This change adds a popen.wait()
to the version checks to insure that the zombies are truly dead.

Review: https://reviewboard.asterisk.org/r/3258/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4775 a5964663-950a-4945-9369-5dd0b5cb08c1