Fix PJSIP AMI show subscriptions test.
A recent change made it invalid to attempt to subscribe to MWI if an
endpoint is already receiving unsolicited MWI. The configuration had
mailboxes configured on both the endpoint an AOR, so the incoming
SUBSCRIBE was not successful.

The change is simply to remove mailbox configuration from the endpoint.
The test now passes.



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4854 a5964663-950a-4945-9369-5dd0b5cb08c1