channel_redirect: wait a bit before hanging up the initiating channel
Strangely enough, in Asterisk 11, it is theoretically possible for a channel
to hit the Hangup application, then let another thread take over for awhile and
run to completion. In this case, that causes this test to fail, as the initiator
channel (which hangs up after the redirected channel) will still emit its event
first. That causes the scenario to cycle, resulting in the Hangup from the
redirected channel in the first scenario to be missed.

A small wait should fix that. Unfortunately, apptest doesn't have a mechanism to
combine events and wait on all expected events before performing an action. Such
a mechanism would be tricky to implement, but if this doesn't suffice, should be
the next step.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4899 a5964663-950a-4945-9369-5dd0b5cb08c1