Allow MWI to be enabled on PJSUA accounts.
The changes to the PJSIP python library pjsua.py that this depends on have
been committed upstream. The test that was up for review with these changes
are actually handled by another test using SIPp. Therefore these changes are
being committed for future test use and the test on the review is not.

Review: https://reviewboard.asterisk.org/r/3348/

(issue ASTERISK-23343)



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4961 a5964663-950a-4945-9369-5dd0b5cb08c1