channels/SIP: Add a test for inbound calls that use TEL URIs
This test verifies support for TEL URIs in basic incoming calls.

An INVITE request with a TEL URI sends a request to Asterisk. The phone-context
specifies a domain for the global number. The From header contains a local
number with a phone-context that contains the prefix of a global number.

If the INVITE request is handled properly, the SIPURIPHONECONTEXT channel
variable will be set properly. If not set properly, the test will fail as the
channel will not be answered and hungup prematurely.

(issue ASTERISK-17179)

https://reviewboard.asterisk.org/r/3441/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4967 a5964663-950a-4945-9369-5dd0b5cb08c1