apps/disa/nominal/authenticate: Fix conflict with extensions.lua
When pbx_lua is loaded with the default extensions.lua, it will add extension
1000@default. Depending on the order in which modules are loaded, this can
cause a 'match more' extension pattern match to believe that it should continue
to receive digits when looking for 100@default. Since the DISA test was
attempting to jump to 100@default, it would fail on the build agents, which
apparently load pbx_lua at a different time than my development machine. Go
Figure.

This patch simply makes the extension that we attempt to jump to 200@default,
which avoids any conflict with the sample extensions.lua extensions.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4973 a5964663-950a-4945-9369-5dd0b5cb08c1