apps/disa/nominal/no_authentication: Add test for DISA that bypasses auth
This test spawns a Local channel that is put into the DISA application. The
application is launched with the 'no-password' option, such that the channel is
not prompted for a password. The channel enters in a valid extension, and the
test verifies that the channel is placed in that extension.

Review: https://reviewboard.asterisk.org/r/3460

ASTERISK-23522 #close
Reported by: Matt Jordan
patches:
  disa_no_authentication.diff uploaded by bford (License 6580)


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@4976 a5964663-950a-4945-9369-5dd0b5cb08c1