fastagi: Fix failing tests by upgrading them to use TestCase
A number of FastAGI tests started failing due to failures in installing their
configuration files. This occurred due to a restructuring of the log file
location. Since this is managed by the TestCase class, it was easier to port
them over to that framework rather than reproduce the logic for managing the
location of the test artifacts.

Note that this isn't a full clean up of these tests, merely a port over to
the correct classes.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5036 a5964663-950a-4945-9369-5dd0b5cb08c1