Increase AMI writetimeout for chatty tests
A number of tests generate a substantial amount of AMI events. When AMI debugging
is cranked up (which it now is if a test produces a 'full' log), the 100 ms write
timeout can be triggered in Asterisk and the test will fail. This patch adds some
manager.users.conf.inc files that bump up the writetimeout for those tests.

git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5040 a5964663-950a-4945-9369-5dd0b5cb08c1