pbx/dialplan: Add an AMI user write timeout
Much like masquerade and a few other tests, the dialplan test is a bit chatty
with AMI user events. As such, given the amount of logging that occurs in
starpy, things can bog down and the connection get broken. This adds a user
level override to increase the write timeout.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5050 a5964663-950a-4945-9369-5dd0b5cb08c1