rest_api/bridges/attended_transfer: Try to start referer scenario second
When using 3pcc, the first scenario that uses sendCmd will establish a TCP
socket connection with the scenario that uses recvCmd first. As such, the
sendCmd scenario *must* start after the recvCmd scenario. The test was,
unfortunately, starting both at the same time. This was highly unlikely to
work reliably.

Unfortunately, knowing when a process has started is a bit tricky. Our SIPp
classes return a deferred that is executed when the process exits; that's not
super useful here. The spawnProcess twisted command returns the IProtocol
instance, which _might_ work, but also doesn't give you easy access to the
state of the process. (You can control the process, but it's not exactly
event driven as to when the process has kicked off). What's more, exposing
the IProtocol instance through our SIPp wrapper would be a bit disruptive.

To work around that, this patch kicks off the referer scenario (which uses
sendCmd first) a few seconds after it kicks off the referee scenario. If
this doesn't prove to be 'good enough', we'll have to find a more
deterministic way to know when the first process is running and the TCP
socket listening.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5061 a5964663-950a-4945-9369-5dd0b5cb08c1