tests/funcs/func_talkdetect: Add a test for the TALK_DETECT function
This adds a test for func_talkdetect. The test consists of four iterations. In
each iteration, the same set of sounds are passed through a Local channel.
TALK_DETECT is placed on one half of the local channel, while playbacks are
initiated via ARI on the other half. Based on the various parameters passed to
TALK_DETECT, a different number of events should be received over AMI/ARI. The
iterations are:
 - An iteration that sets TALK_DETECT with its default settings. A single pair
   of events is expected to be received.
 - An iteration that sets TALK_DETECT with a short silence threshold and a
   different talk threshold. Two pairs of events are expected to be received.
 - An iteration that sets TALK_DETECT with a very short silence threshold.
   Three pairs of events are expected to be received. 
 - An iteration that removes TALK_DETECT. No events should be received.

The test verifies that a total of 6 event pairs are received (over both
AMI/ARI), and that in the 'end' events, a duration that is non-zero is present.

Review: https://reviewboard.asterisk.org/r/3564/

ASTERISK-23786
Reported by: Matt Jordan


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5085 a5964663-950a-4945-9369-5dd0b5cb08c1