Add an RLMI module and demonstration script.
RLMI is an XML document type used for resource list subscriptions.
In the near future, tests will be written that will need to take
deep looks at the content of RLMI documents. To make this easier,
I have use pyxb [1] in order to create an RLMI XML parser. This
parser, rlmi.py lives in the testsuite's lib/python directory.

In addition, I have added an rlmi_demo.py script to the contrib/scripts
directory to illustrate how to use rlmi.py to browse the elements of
the RLMI XML document.

[1] http://pyxb.sourceforge.nethttp://pyxb.sourceforge.net


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5112 a5964663-950a-4945-9369-5dd0b5cb08c1