manager: Add a status action and channelvars tests
This patch adds tests to cover code affected by the proposed change on review
https://reviewboard.asterisk.org/r/3568.

This patch includes tests for AMI Status and the AMI channelvars setting. This
action and that setting were affected by the patch. Both tests pass with and
without the presence of the patch.

Review: https://reviewboard.asterisk.org/r/3578/

ASTERISK-23811


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5120 a5964663-950a-4945-9369-5dd0b5cb08c1