test_case: Revert AMI shutdown patch
Unfortunately, there are a lot of tests that do not have a good, clean shutdown
sequence. These tests are typically still looking for an AMI event during the
shutdown sequence; closing the AMI connection prior to issuing the shutdown
causes them to fail.

While this patch is the right solution long term, the tests should be cleaned
up prior to this being applied. As it is, the large number of test failures
resulting from this patch make it more difficult to run the build agents. Given
the other integration efforts currently ongoing, we'll have to revisit this
patch at a latter date.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5274 a5964663-950a-4945-9369-5dd0b5cb08c1