testsuite: fail sipp test if sipp not installed
If sipp is not listed as a dependency and isn't
installed, the sipp test case hangs while it is
trying to kill a nonexistant process.  Instead,
check if it is installed and raise an exception
if not.  Also fixed the missing dependancy that
precipitated this discovery.

Review: https://reviewboard.asterisk.org/r/3884



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5341 a5964663-950a-4945-9369-5dd0b5cb08c1