apptest: Fix issue where duplicate event matches will re-run actions
When the action handling was modified for the inclusion of the off-nominal
authenticate test, a bug was introduced where a duplicate AMI event will
cause the same actions to be executed again. Previously, actions were
not removed from the action list and a counter was maintained that specified
where in the action lsit we were executing. Post the off-nominal test,
actions are removed from the list as they are executed. Unfortunately, the
actions on the list were copied, resulting in the original list being
maintained. This only caused an issue when the event matcher matched multiple
events; however, when that happened, the same actions would be executed.

This caused the confbridge triple lindy test to fail, as it spawned multiple
channels that were not expected.

Along the way, this patch also makes a small improvement to matching
application and candidate channels. Prior to this patch, application and
candidate channels could be tied together in a slightly more loose fashion;
now those channels must be either side of a Local channel pair. This fixes a
bug that was noticed when the triple lindy test was being debugged, wherein
the wrong application channel was paired with a controller channel.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5463 a5964663-950a-4945-9369-5dd0b5cb08c1