lib/python/asterisk: Fix comparison function for large size differences
In r5615, a comparison function was introduced that attempts to pick the
directory with the most space available, comaparing /tmp and /var/tmp.
Unfortunately, if the blocks available are substantially different, the
value computed by the comparison function is a long, which causes a run-time
exception. Comparison functions must return an integer.

As such, this patch updates the comparison function to always return -1, 0,
or 1. If those aren't integers, then we might as well pack it all up and go
home.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5627 a5964663-950a-4945-9369-5dd0b5cb08c1