Fix 'expected-result' YAML property & santize return code from test_runner.py
This fixes the 'expected-result' YAML property for test configurations. When
it is set to False and the test fails, the test will properly be marked as
passed. When is is set to False and the test passes, the test will properly be
marked as failed since it didn't match what was expected.

Additionally this sanitizes the return code from test_runner.py so that
'self.pass' is always a boolean whereas before there were some scenarios that
it would be an int.

Review: https://reviewboard.asterisk.org/r/4080/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5767 a5964663-950a-4945-9369-5dd0b5cb08c1