Add test to unregister registered outbound registrations.
This tests unregistering registered outbound registrations using the
PJSIPUnregister AMI action. It ensures that Asterisk sends a REGISTER message
with an Expires header of '0' upon executing the AMI action. It also ensures
that AMI 'Registry' events are present with a status of 'Unregistered'. All
combintations of IPv4/IPv6 and UDP/TCP are utilized.

Notes:
* The 'expected-result' YAML property is set to False as the test is
currently expected to fail due to bug ASTERISK-24411. The property should be
removed when the bug is resolved.
* The SIPp scenario contains a pause at the end to work around bug
ASTERISK-24414. It should be removed when the bug is resolved.

Review: https://reviewboard.asterisk.org/r/4077/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5773 a5964663-950a-4945-9369-5dd0b5cb08c1