testsuite: avoid lstrip when removing prefix
This fixes a problem where the relative path
to the test would be missing some characters
if they matched t, e, or s. The new value of
test_relpath now provides a correct path for
the rest of the TestRun class.  Also fixed a
similar case in the sipp version check.

Review: https://reviewboard.asterisk.org/r/4083/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5804 a5964663-950a-4945-9369-5dd0b5cb08c1