tests/rest_api/channels/originate_then_continue: Fix stability problems
In r5818, the ARI test object was modified to close the WebSocket on test
completion. This helps with reference leak detection, as it causes Asterisk
to clean up resources prior to exiting. Unfortunately, a number of tests were
written in a fashion that assumed events would continue to be received even
after the test was told to stop. This test - originate_then_continue - was
one of them.

A large part of the problem is that the ARITestObject (and its derivatives)
will stop the test when all test iterations have completed. That's fine when
the test iterations - and the events that occur within them - comprise the
entire test. If a test has events outside of that, the test is highly likely
to fail, as stopping the test now closes the WebSocket.

This updates the test object with a new parameter, stop-on-end, that allows
a test to not end the test when all iterations are done. The test - using
pluggable modules - can then stop the test when it wants.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5839 a5964663-950a-4945-9369-5dd0b5cb08c1