rest_api/bridges/{hangup|happy}: Fix test stability issues
This patch makes the tests stop when the last expected event is received, and
not when the test iteration completes in the underlying test object. This
prevents the WebSocket being closed by the call to stop_reactor from causing
events to be missed.

How many different ways can I type this commit message? :-)


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5863 a5964663-950a-4945-9369-5dd0b5cb08c1