testsute: Add 'add-relative-to-search-path' handler to TestModuleLoader
I've got some common code I want to share between a few tests so it's
in the test's parent directory.  However, there doesn't seem to be a way
to add the parent directory to the search path without specifying an
absolute path.  So, I added a helper to TestModuleLoader that will allow
you to add paths relative to the test directory.

Example:

test-modules:
    add-test-to-search-path: 'True'
    add-relative-to-search-path: ['..']
    test-object:
        config-section: object-config
        typename: 'ManagerConfigTest.ManagerConfigTest'
                            
This adds <test_path>/.. to the search path. I.E. The test's parent directory.

Tested-by: George Joseph
Review: https://reviewboard.asterisk.org/r/4134/




git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5894 a5964663-950a-4945-9369-5dd0b5cb08c1