runtests.py: Don't completely bail on the script if unable to kill processes
Exiting the script completely when unable to kill Asterisk or SIPp is bad for
a few reasons:
(1) We get little to no output from the build agents when this occurs.
(2) os.system's return value is system dependent. Hence, this could return
    something other than 0 on success.
(3) Even if we can't kill a runaway Asterisk or SIPp process, that is no
    guarantee that subsequent tests will fail. Attempting to purge them before
    a run is a good thing; giving up if that doesn't work is not necessarily
    better than trying anyway.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5963 a5964663-950a-4945-9369-5dd0b5cb08c1