Add test for transfer retries option.
This adds a test that performs two scenarios:
1) The transferer attempts a transfer but dials the incorrect
extension. The transferer is then given another attempt to dial
the correct extension and dials the correct one this time. Ensure
the transfer succeeds.
2) The FEATURE() dialplan function sets the number of attempts
for the transfer to 1. The transferer attempts a transfer but dials
the incorrect extension. The transfer attempt should fail at this
point since no retries are permitted.

Review: https://reviewboard.asterisk.org/r/4168


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5985 a5964663-950a-4945-9369-5dd0b5cb08c1