asterisk: Fix timeout issues introduced by r5942
When the valgrind patch was merged for r5942, it unilaterally increased the
timeouts for the callback sequence during an Asterisk shutdown. When we
shutdown Asterisk, we:
* Call stop gracefully
* Wait a bit, then call stop now
* Wait a bit, then kill Asterisk

The wait times for killing Asterisk was increased from 5 seconds to 200
seconds. For tests that currently KILL Asterisk, this made them run
inordinately long. It also increased unilaterally the 'core wait fullybooted'
timeout.

This patch restores the regular timeouts when VALGRIND_ENABLED is not detected.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5993 a5964663-950a-4945-9369-5dd0b5cb08c1