bridge/atxfer_retries: Unilaterally kill all channels on scenario end
When the atxfer_retries test ends, it does properly hangup the 'test' channels
that it controls. However, because it dials several other channels in the test,
and those channels are transferred to different extensions, there are several
channels still active and not controlled by the AppTest framework active. When
Asterisk is shut down gracefully, the graceful shutdown fails and we end up
having to kill Asterisk.

This patch updates the test to send a 'Hangup all the channels' AMI action
just prior to scenario end. That causes all of the transferred channels to be
hung up as well.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@5994 a5964663-950a-4945-9369-5dd0b5cb08c1