rest_api/mailbox/baseline: Remove test for StasisEnd event
The ARI base test object ends a test when it detects a Hangup over AMI. This
is not guaranteed to occur after a StasisEnd event. Since the StasisEnd event
has no bearing on the intent of this test, this patch removes it to prevent
a race condition where Asterisk is shut down before the test case has the
opportunity to process the event.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6001 a5964663-950a-4945-9369-5dd0b5cb08c1