tests/bridge/atxfer_retries: Change context for test to avoid conflicts
The atxfer_retries test relies on the fact that certain extensions not in its
dialplan - such as 1235 - will not exist. Unfortunately, the build agents
use 'make samples' - which destroys any notion of won't exist in a dialplan.
In this case, pbx_lua provides a 1235 extension in context 'default', which
caused the transfer that was *supposed* to fail to succeed.

This patch simply changes the context the test uses from 'default' to
'atxfer_retries'. That should prevent any false matching during the attended
transfers that are supposed to fail.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6019 a5964663-950a-4945-9369-5dd0b5cb08c1