rest_api/applications/double-subscribe-device-state: Fix test stop race
The double-subscribe-device-state test uses a Local channel to trigger the test
logic. The default behaviour of the test object is to stop the reactor when the
channel hangs up. However, there is no guarantee that the device state changes
have all propagated and the ARI events been received. This patch makes it so
that the test object doesn't stop the reactor on channel hang up - instead, the
handling of the device state triggers the reactor stop.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6026 a5964663-950a-4945-9369-5dd0b5cb08c1