channels/pjsip/transfers/blind_transfer: Fix refer only premature transfer
The callee_refer_only|caller_refer_only tests use PJSUA Python bindings to
initiate a blind transfer of the opposing party. Prior to this patch, the
test would initiate a transfer when it detected that the initiator of the
transfer was up and in a call. However, this does not always mean that the
party is ready to initiate the transfer: in the case of the callee, they
will be answered prior to the caller; in the case of the caller, they may
be answered but not fully in the bridge.

This patch makes it so that the initiator of the transfer does not perform
their action until we receive BridgeEnter events for both of the original
parties. This prevents the Blind Transfer from failing due to not having
the initiator in a bridge with another party.

ASTERISK-24574
Reported by: Joshau Colp


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6033 a5964663-950a-4945-9369-5dd0b5cb08c1