pjsip: Add a test which performs a hold and includes a declined media stream.
This test establishes a perfectly normal call with audio and a declined video media stream.
Once established the call is put on hold, including the declined video media stream in
the SDP. Asterisk should respond with a 200 OK to this situation and not a 488.

ASTERISK-24607 #close
Reported by: Matt Jordan

Review: https://reviewboard.asterisk.org/r/4254/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6083 a5964663-950a-4945-9369-5dd0b5cb08c1