testsuite: Allow tests to specify multiple minimum versions
Prior to this patch, a test could only provide a single required minimum
version of Asterisk. This didn't bode well for tests that have been added due
to bugs, which typically require a minimum version for each active branch of
Asterisk. Not allowing multiple minimum versions to be specified would result
in the test being erroneously executed for higher branches of Asterisk,
resulting in false test failures.

Take, for example, tests/channels/SIP/no_ack_dialog_cleanup. This test was
added for ASTERISK-20784, the fix of which was released in 1.8.32.0, 11.14.0,
12.7.0, and 13.0.0-beta3. Because only a single version could be specified,
however, the best the test writer could do was provide the 1.8 version:

properties:
   minversion: '1.8.32.0'

This would prevent 1.8.31.0 from running against the test, but 11.13.0 would
execute the test, as it is greater than the supplied minversion of 1.8.32.0.

However, if we specify a list of minversions:

  minversion: ['1.8.32.0', '11.14.0', '12.7.0', '13.0.0-beta3']

Then we can determine that 11.13.0 is less than 11.14.0, and prevent the test
from running.

With this patch, that is now possible. Failed Bamboo runs rejoiced everywhere.

Review: https://reviewboard.asterisk.org/r/4270/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6127 a5964663-950a-4945-9369-5dd0b5cb08c1