Add a test for PJSIP keep alive packets for connection oriented transports
This patch adds a test for the PJSIP keep_alive_interval option. It sets the
keep_alive_interval to 2 seconds, makes a TCP client connection to the PJSIP
stack, and then waits for 5 keep alives. Upon receiving 5 packets, it
disconnects and stops the test.

In all cases, the test verifies on stopping:
* That 5 total packets were received
* That the time deltas between packets were received within approximately 2
  seconds of each other
* That the keep alive values are all '\r\n\r\n'

Review: https://reviewboard.asterisk.org/r/4293/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6134 a5964663-950a-4945-9369-5dd0b5cb08c1