Add a test for user_eq_phone setting in PJSIP
This patch adds a test for the user_eq_phone endpoint setting in PJSIP.

The test verifies that when the user_eq_phone setting is enabled on a PJSIP
endpoint, a request sent from Asterisk to that endpoint that contains a
telephone number in the request URI has a 'user=phone' specified appended to
it. The test originates a Local channel that causes an outbound dial to number
12568675309 at endpoint 'jenny'. The SIPp scenario verifies that a 'user=phone'
tag is found in the INVITE request received from Asterisk.

Review: https://reviewboard.asterisk.org/r/4294/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6135 a5964663-950a-4945-9369-5dd0b5cb08c1