channels/pjsip/keep_alive: Fix format issue for Python 2.6-
In Python 2.6-, the string format method requires positional arguments. Not
providing them (which is allowed in 2.7+) results in an exception. This patch
corrects the log messages with positional arguments, which is valid for all
reasonable versions of Python.

On another note, can we all agree that Python 2.7 should be the defacto
2.x standard on CentOS by now? I was shocked when I saw that CentOS 6 delivers
2.6 by default. Sheesh.

And... thus ends my last commit in 2014. Here's to 2015!


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6157 a5964663-950a-4945-9369-5dd0b5cb08c1