Add tests for continuation and origination.
There was already a test for continuation, but this expands the test to cover many more
test cases.

This also adds a test originating calls to different dialplan locations to make sure that
different combinations of extension, context, priority, and label work as expected.

Review https://reviewboard.asterisk.org/r/4284



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6192 a5964663-950a-4945-9369-5dd0b5cb08c1