testsuite: check for channel leak on failed blonde transfer
Tests attended transfer converted to blind by hangup of
the transferrer (blonde transfer), but where the target
of the transfer is then later unavailable, which causes
the transfer to fail and attempt to recall the original
transferer.   After the recall is answered, the list of
active channels is checked for a leaked reference.

ASTERISK-24513
Review: https://reviewboard.asterisk.org/r/4256/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6227 a5964663-950a-4945-9369-5dd0b5cb08c1