Testsuite: Created tests to ensure that ARI behaves as expected when updating a bridge through the /bridges/{bridgeId} endpoint.
The tests exercise all expected combinations where posting to the /bridges/{bridgeId} enpoint should fail. If a bridge with the given <bridgeId> already exists in Asterisk, the post is expected to be treated as an update. The expected behavior is to reject requests that include any value for the query parameter <type> and/or a value for the query parameter <name> that differs than the name used when creating the bridge, and return an HTTP status code of 500 (Internal Server Error) upon rejection of the request.

ASTERISK-24560 #close
Reported By: Kinsey Moore

Review: https://reviewboard.asterisk.org/r/4348/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6319 a5964663-950a-4945-9369-5dd0b5cb08c1