apps/dial/mid_call_events/dial_{gosub|macro}: Verify off-nominal return values
This patch updates the mid call events tests to verify that mid-call
Macros/GoSubs whose MACRO_RESULT/GOSUB_RESULT is set to an unknown value does
not result in multiple DialEnd events. The test spawns an additional channel
that calls a mid-call Macro/GoSub that returns, in its variable, a value that
the Dial application does not understand. The test verifies that the two
channels are bridged as expected, and that only a single DialEnd event is
received.



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6333 a5964663-950a-4945-9369-5dd0b5cb08c1