Add test to ensure no channel is leaked if we receive a REINVITE from a channel that performs a transfer.
This test has a SIPp scenario perform a blind transfer that fails. Upon receiving the failure notification,
the SIPp scenario sends a reinvite to Asterisk. The test ensures that this reinvite does not cause the
creation of a new channel by ensuring that the channel count after the test completes is zero.

Review: https://reviewboard.asterisk.org/r/4343



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6364 a5964663-950a-4945-9369-5dd0b5cb08c1