tests/channels/SIP/sip_attended_transfer: Make test a bit less bouncy
Depending on the speed of the machine, we may get a few extra Link/Unlink
Bridge events. Due to how the test was tracking the channels, this could
cause it to think that the same channel was bridged with itself. As a result,
we'd fail to detect things later in the test. This patch updates the test to
not track the same channel twice.

This should make the test more reliable on versions prior to Asterisk 12.
There's still a few other test failures in the Test Suite, wherein PJSIP
fails to return control to the Test Suite while waiting for registration.
This is due to it conflicting with twisted - but that's a problem for another
patch.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6432 a5964663-950a-4945-9369-5dd0b5cb08c1