tests/apps/confbridge/confbridge_recording: Handle race conditions in recording
When recording a conference, there is a race condition between the MixMonitor
on the recording channel being finished before/after the conference finishes.
This can cause the test to fail during the sound checking portion.

This patch does the following:
(1) It makes it so that the end of the conference logic (wihch triggers the
    sound detection) is fired either from the MixMonitor end logic or the
    hangup of the channel in the ConfBridge. In both cases, the logic checks
    to see if the other condition already occurred, and only triggers the
    sound checking logic if the other one did occur.
(2) The sound checking logic is now a bit more careful about not firing twice,
    as that can cause the test logic itself to execute more times than it
    should.



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6442 a5964663-950a-4945-9369-5dd0b5cb08c1