tests/channels/SIP/sip_blind_transfer: Start PJSUA instances earlier
In some test failures for the sip_blind_transfer tests, Asterisk will receive
a 503 back from a PJSUA instance that it sent an INVITE request to. This is
most likely due to the PJSUA instance not yet being fully initialized and ready
for an inbound INVITE request. The PJSUA instances used for these tests were
originally instantiated during the AMI connect callback; this meant that the
processes were spawned almost immediately before the calls were made from
the callee instance to Asterisk to the receiver instance.

This patch moves the startup of the PJSUA instances to the TestCase derived
class's constructor. This gives them a bit more time to spin up, as Asterisk
itself is not yet created at that point. This should make the tests slightly
more reliable.

Note that these tests could stand some serious refactoring to make use of the
new phone controller module that was just merged, but that's a patch for
another day.


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6451 a5964663-950a-4945-9369-5dd0b5cb08c1