testsuite: Add DNS server pluggable module.
This change adds a pluggable module to the testsuite which
runs a DNS server using twisted-names. This is an authoritative
DNS server that responds with records which are specified in a
Python style zone file or a BIND style zone file. The DNS server
lives for the lifetime of the test. This is useful since it removes
any reliance on an external DNS server and reduces the work required
to test DNS.

Review: https://reviewboard.asterisk.org/r/4451/


git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6467 a5964663-950a-4945-9369-5dd0b5cb08c1