Testsuite: Generate UserEvent upon PJsua phones being ready.
This adds the generation of a UserEvent on all AMI connections when all PJsua
phones are registered and ready. Originating calls to the PJsua phones should
be triggered by the 'PJsuaPhonesReady' UserEvent instead of the registration
event or any others to prevent tests from bouncing. The test has been changed
to use this event which prevents it from randomly failing.

This also adds a sample YAML config for using the phones module.

Review: https://reviewboard.asterisk.org/r/4470/



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6523 a5964663-950a-4945-9369-5dd0b5cb08c1