pjsua doesn't send a 180 ringing
Previously the 'update_connected_line_information' routine was always sending a
180 RING, but after some changes were made it will only send one if it needs to.
Apparently, pjsua not receiving a 180 caused it to later send an out of dialog
REFER instead of an in dialog one. This of course is not handled by Asterisk.
Since this seems to be a problem in pjsua for now this change makes sure a ring
is always sent via the dialplan, thus pjsua will now send the REFER in dialog
and the test passes.



git-svn-id: https://origsvn.digium.com/svn/testsuite/asterisk/trunk@6540 a5964663-950a-4945-9369-5dd0b5cb08c1