PJSIP: Added test to ensure retransmissions are not handled.
In this test, a SIPp scenario sends the exact same MESSAGE request
to Asterisk twice. The test ensures that the dialplan is only called
into a single time.

Without the patch from https://reviewboard.asterisk.org/r/4532/ , this
test fails because the UserEvent in the dialplan is sent twice. With
that patch, this test succeeds.

ASTERSIK-24920
Reported by Mark Michelson

Change-Id: I524a3eb1cde4489d0ff9866913ae1be318c72115