tests/rest_api/channels: Add a channel hold intercept test
This test verifies that a channel that initiates a call hold with the
HOLD_INTERCEPT function enabled on it will have the hold frame intercepte
and turned into an event for ARI clients.

A Local channel enters the Stasis application, and the HOLD_INTERCEPT function
is placed on the channel. A POST /hold operation is then used to simulate a
call hold on the channel. The test verifies that a ChannelHold event is raised
with the channel initating the hold. A DELETE /hold operation is then used, and
a ChannelUnhold event is raised with the channel initiating the unhold.

ASTERISK-24922
Reported by: Matt Jordan

Change-Id: If8bdbc74e0789d0c13ffbc5cb0c852f5f3a04012