tests/rest_api/channels/hold: Add a default Hold action test
This test verifies that a channel that initiates a call hold will have the
Hold event raised for it, and that a channel that removes a call hold will
have an Unhold event raised for it.

The test puts two Local channels into a Bridge together. When both channels
are in a bridge, an ARI POST /hold operation is performed on the first
channel. When the Hold event is received, an ARI DELETE /hold operation is
performed on the same channel. When the Unhold event is received, the
channels are removed from the bridge and hung up, and the bridge is destroyed.

ASTERISK-24922
Reported by: Matt Jordan

Change-Id: I9786cb0829e39b33a8dbc130e53fd15a1b5e0d68