Rewrite sip_attended_transfer test to stop failing.
The sip_attended_transfer test has been bouncing for a while. There
are two major fixes introduced here to prevent the bouncing.

First, by converting to using the testsuite's PJSUA module, we no
longer are using the native Python threading library. Instead, we
are using a method that works better with the Twisted framework.

Second, the test is more strict about when the transfer may be
performed. The previous test would attempt the transfer when Asterisk
reported that the call was bridged. The problem is that Asterisk may
report the call as bridged before PJSUA has properly processed the
200 OK that Asterisk has sent to it. By waiting, we can be sure that
all parties are prepared when the transfer is attempted.

The test has also been rewritten to only work with Asterisk 12+. A
new separate test will be written to work on Asterisk 11. This helps
the code to be a little less cluttered.

Change-Id: I1676801d90bcafc28ba25e8b6889f40ab08cc90e

Address review feedback from Ashley

* Two CallCallback classes have been combined into one
* Bridge state has been factored into a minimal class
* Unnecessary checks of test state have been removed.

Change-Id: I1676801d90bcafc28ba25e8b6889f40ab08cc90e