Add SIP attended transfer for Asterisk 11.
The structure of this is similar to the new Asterisk 12+ SIP attended
transfer test, except that the method of determining if channels are
bridged is different.

Like the 12+ version, the two major changes to the original test are
the use of the PJSUA testsuite module and waiting to move forward
with test actions until PJSUA has informed us of state changes.

The logic used for determining bridgedness is the same as was used
in the sip_attended_transfer test that is being replaced.

Change-Id: I48c7b6a9298552aa756d0c2f26afbd6a96d553b5