rest_api/channels/snoop_spy: Stop test on bridge destruction
The snoop_spy test is currently bouncing on the CI build agents. This is due to
a race condition in the test:
* The test wants to verify that three StasisEnd events have occurred. When all
  events have occurred, it will destroy the bridge to clean up resources.
* At the same time, the twisted reactor will be stopped by the test object when
  the Local channel it spawned is hung up.

This causes a race condition, as the Local channel hangup is not ordered with
respect to the StasisEnd events received by the test. As a result, the
WebSocket connection may be broken before all StasisEnd events have been
received.

This patch fixes this by:
1. Instructing the test object to not stop the test when the Local channel has
   been destroyed.
2. Introducing a new instance of a pluggable module to watch for the
   'BridgeDestroyed' event. Since there is only one bridge in the test being
   destroyed at the very end of the test, this is easy to watch for.
3. When said event is received, stop the test.

This guarantees that the test will finish to completion, and all resources are
reclaimed prior to stopping Asterisk.

Change-Id: I373981f81ca455743bbf2371f07b380d013cd12b