Testsuite: New test for FAX via PJSIP T38 with authentication
Add a test for PJSIP t38 with authentication based on normal t38 test
The test will start two instances of Asterisk. The first will originate
a PJSIP call with authentication to the second using an extension that
will run sendFax. The second will receive the call and direct it to an
extension that runs receiveFax. As the fax completes, user events are
issued to verify that the FAX completed successfully.

ASTERISK-24933
Reported by: Jonathan Rose
Change-Id: If37cf20857ae3c0b35e0637a0a2cb7e7d6226df6