non_stasis_bridge_to_stasis_bridge: Update regex for ami events
Due to a bug in Asterisk, in some cases, the channel would not be hung up after
leaving the bridge upon completing an attended transfer. With that fixed this
test can now be expected to pass. However there were a few spots where the ami
event matching class had trouble matching events with a '-' included. Removed
the '-' to no longer be part of the search criteria. The test now works as
expected.

Change-Id: I0a3bcb1a0df7e7bdca02be827288f5f08b5140ce