non_stasis_bridge_to_stasis_bridge: Update regex for ami events
Due to a bug in Asterisk, in some cases, the channel would not be hung up after
leaving the bridge upon completing an attended transfer. With that fixed this
test can now be expected to pass, so the 'expected-result: False' has been
removed. Also a couple of events attempted to match against the wrong regex,
so those were modified accordingly.

Change-Id: I0a3bcb1a0df7e7bdca02be827288f5f08b5140ce