pjsip/dtmf_incompatible: Wait 1 second before hanging up Alice
The dtmf_incompatible test bounced on one of the CI build agents due to
only receiving a single BridgeLeave event. This was because Alice never
made it into the bridge: the BYE request was received from the SIPp
scenario prior to her channel fully entering the bridge.

This patch simply puts an additional <pause> into the SIPp scenario that
waits for one second before sending the BYE request. While waits and
pauses are generally not ideal, it is reasonable to expect that a
channel will make it into a bridge after one second. If not, we probably
have larger concerns.

Change-Id: Ia2a5f42255019f7fef2aa3d6cba1bce5123329d8