python/asterisk/version: Update version handling for Git migration
This patch updates the Asterisk version parsing module to handle
versions derived from a Git repository. It leaves SVN and tag version
parsing alone.

All Git versions are assumed to be in the following form:
  Asterisk GIT-{branch}-{obj_blob}
Where branch must be 'master' or one of the canonical mainline branches.

Since the version tag no longer reflects individual team or local
branches, this makes it far easier to map a team/local branch to an
upstream Asterisk version.

ASTERISK-24954

Change-Id: Idbb0ef92fec6d4e8e39da105c9c635f9ecdd4ee4