non_stasis_bridge_to_stasis_bridge: hangup race
During the transfer process there is a chance for a hang up to occur on the
channel being transfered (in this case bob) but before the transfer has time
to fully complete. If this happens not all expected events are raised, thus
not received. Addressed the possibility of this scenario occurring by adjusting
the expected events in question to accept at most one of the event, consequently
allowing zero as well.

Change-Id: Ic4ef9fff4830670137dbfb587f30eaf33a43a2d6