testsuite: Remove some external dependencies for run-local
run-local wasn't accounting for 64 bit systems where libraries are in
lib64 instead of lib.  Now it searches for libasteriskssl.so in ./astroot
and prepends it's path to LD_LIBRARY_PATH.

buildoptions.py now includins ./astroot/usr/include/asterisk when
looking for buildopts.h.

When run with run-local the testsuite should now not look for any
asterisk components outside of ./astroot.

ASTERISK-24984 #close

Change-Id: Idf59b8b6de7524913ab4e3c29006e91a29d60e4e