Fix failing masquerade test on 13+
The 13 branch recently received a change that by default prevents a
chain of more than 20 channels from being dialed on a single call. This
causes the masquerade test to fail since it expects to have a chain of
200 channels to optimize.

The fix is to use the CHANNEL(max_forwards) function to allow for enough
of a chain to be created. With this patch, the failing masquerade test
now passes.

ASTERISK-24985 #close
Reported by Mark Michelson

Change-Id: I87d76a0bb245b977bb0b2e545e6f38088ddb2f00