pjsip: Add res_resolver_unbound as a dependency for resolver tests.
Since the PJSIP resolver tests are testing new resolver support this
requires the res_resolver_unbound module. Without it these tests will
fail. In the future when a fallback resolver implementation is
available this requirement can be removed, but for now it is required.

ASTERISK-24987 #close
Reported by: Joshua Colp

Change-Id: I9ce4d94223122b35f8e6d871151904b3c02b389b