Testsuite Test 'stasisstatus' Fails During Build
During execution of the test, the first test scenario, [Babs], fails. According
to the logs, the problem begins when the web socket for [Babs] fails to
instantiate on the first try. While it is attempting to reconnect, the TestCase
starts the next TestScenario, [Bugs], and leaves [Babs] orphaned. This hang up
subsequently results in the twisted.reactor timing out. The logic in the test
states if the reactor times out, the test is automatically marked as a failure
(as it should).

This patch attempts to remedy the above situation by adding an observable
event, 'on_start' for the TestScenario class. The observers for this event are
notified when the AriClient has completed creating the web socket. The TestCase
registers itself as an observer of this event and then uses its callback as a
means of scheduling the start of each subsequent TestScenario.

ASTERISK-24995 #close
Reported By: Ashley Sanders

Change-Id: Iad3d6b3a092d8f57dfd7855c79b641096de129aa