Add local attended transfer tests with direct media, hold, and REFER/Replaces.
This adds two pjsip tests for caller and callee initiated local attended
transfers with direct media, hold, and REFER/Replaces. Both tests use the
packet_sniffer.py module to check AMI UserEvents that it generates.

* packet_sniffer.py: This sniffs SIP packets to track calls. It generates an
AMI UserEvent when a remote bridge is setup on two call legs, when a remote
bridge is torn down on a call leg, and when a REFER, Notify sipfrag, and 491
Request Pending is found.
* lib/python/asterisk/pcap.py: Fix the finding of SIP packets and attribute
references.
* lib/python/asterisk/phones.py: When a pjsua operation (transfer/hold) is
performed and before any callbacks are called or state changes occur, there is
no way to know that an operation is in progress. This adds tracking of calls
with operations (transfer & hold) that are in progress. When an operation is in
progress for a call, this prevents the same operation from being performed on
the call. This also adds additional call & media state checking and updates
some of the exception handling.

ASTERISK-23644 #close

Change-Id: Ibd1f9a3ea8bf6b80fe95efe3fcd880a55d5a7cae