pjsip: Add resolver test for SRV priority with UDP transport.
This test does a SRV record lookup resulting in two records each with a
different priority and port for the UDP transport. An IPv4 & IPv6 address (A &
AAAA records) is provided for each SRV record. Port 5061 is used if the highest
priority SRV record (lowest number) is chosen and port 5062 otherwise. If both
IPv4 & IPv6 calls do not reach port 5061 (where SIPp is listening on) then the
test fails.

ASTERISK-25009 #close

Change-Id: I8efdf78580cd327452a22863622521fd1a9385bb