runtests.py: Add parameter --random-order.
This adds a parameter to runtests.py that allows the tests to be run
in a randomized order instead of alphabetically.  This can be useful
when testing a change to the Asterisk that requires running the full
testsuite.  For example if the change breaks rest_api, you will not
have to wait for 80% of the testsuite to run before those tests start
to produce failures.

Change-Id: I002c6fa2c7a05e44673cb56b11d4e4c381f3eba0