Expand checks for core dumps.
Tests using lua (asttest) change to the test directory.  This causes
any core dumps from Asterisk to be placed there instead of the
testsuite root directory as expected.  Running with the -V option
produces a file named vgcore.${PID}.  These variations of name and
location are now recognized and processed.

Change-Id: Idf83bff672cb4c3559214864498a9baf404c1fc5