Add a nominal caller initiated remote attended transfer test.
This uses two Asterisk instances and a PJSUA instance. The PJSUA
instance makes a call to both Asterisk instances and then performs an
attended transfer. This ensures that the caller initiated remote
attended transfer properly occurs.

The pjsua_mod.py module had to be modified to allow the creation of
PJSUA accounts that do not register. Otherwise PJSUA will only send
calls to the registrar that it is registered to even if the IP of the
dialed URI differs. Thus the test configuration option 'register' was
created to allow disabling registration for PJSUA accounts. If
registration is disabled, the callback module for user code is called
after all accounts have been created.

The phones.py module was updated to override pjsua_mod.acct_success()
that has been added. This is so the AMI PJsuaPhonesReady user event is
generated when registration is disabled and all accounts have been
created.

Change-Id: I5aaf2b4a72d3334e320c7056a4eed8176b71ce6e