Enable retention of process maps when it might be useful.
This change has no purpose for current versions of Asterisk.
They do not call dlclose during shutdown, so backtraces from
the testsuite normally should not contain unresolvable symbols.

This should be used for all testing of the module loader rewrite:
https://gerrit.asterisk.org/557

Final Commit Message:
---------------------
Make a copy of /proc/$PID/maps for each Asterisk process when it
is fully booted.  This file is then copied to testsuite logs if
valgrind is enabled or if core dumps were detected.

The process maps file is only copied for Asterisk 14 or above.
Previous versions of Asterisk never dlclosed modules during the
testsuite, so the maps file is unneeded.  The file is only copied
when it could be useful since it can easily be 400k.  Even without
core dumps or valgrind this can add up very quickly if tmp is not
cleaned up after every test.

A find_symbol.py script is added to allow retrieving symbol
information using a proc-maps.txt and an unresolvable pointer
address from the backtrace.

Change-Id: Ie67750892cbe0e754061c90adedf682840b069b0